+++ NEWS.md
@@ -1,6 +1,7 @@
 # surveydown (development version)
 
 - `sd_redirect()` to create redirection to external links, either by providing a button, a countdown, or both.
+- Now `sd_next()` and `sd_redirect()` both support the "Enter" key for a better user experience.
 
 # surveydown 0.2.1